+++ tests/decisions/test_decisions.py
@@ -1,12 +1,13 @@
+from pathlib import Path
 import pytest
 
 
-from tests.conftest import IS_LINUX, GcovrTestExec
+from tests.conftest import IS_LINUX, USE_PROFDATA_POSSIBLE, GcovrTestExec
 
 
 @pytest.mark.skipif(
     not IS_LINUX,
-    reason="Parsing of decision is independent of OS and we do not want to have separate data wor Windows and Darwin.",
+    reason="Parsing of decision is independent of OS and we do not want to have separate data for Windows and Darwin.",
 )
 def test_decisions(gcovr_test_exec: "GcovrTestExec") -> None:
     """Test of decision parsing."""
@@ -22,20 +23,17 @@ def test_decisions(gcovr_test_exec: "GcovrTestExec") -> None:
         "--verbose",
         "--decisions",
         "--json-pretty",
-        "--json",
-        "coverage.json.gz",
+        "--json=coverage.json.gz",
     )
     gcovr_test_exec.gcovr(
         "--verbose",
-        "--json-add-tracefile",
-        "coverage.json.gz",
+        "--json-add-tracefile=coverage.json.gz",
         "--json-pretty",
         "--json=coverage.json",
     )
     gcovr_test_exec.gcovr(
         "--verbose",
-        "--json-add-tracefile",
-        "coverage.json.gz",
+        "--json-add-tracefile=coverage.json.gz",
         "--decision",
         "--json-summary-pretty",
         "--json-summary=coverage_summary.json",
@@ -44,8 +42,7 @@ def test_decisions(gcovr_test_exec: "GcovrTestExec") -> None:
 
     gcovr_test_exec.gcovr(
         "--verbose",
-        "--json-add-tracefile",
-        "coverage.json.gz",
+        "--json-add-tracefile=coverage.json.gz",
         "--decision",
         "--html-details=coverage.html",
     )
@@ -53,10 +50,8 @@ def test_decisions(gcovr_test_exec: "GcovrTestExec") -> None:
 
     process = gcovr_test_exec.gcovr(
         "--verbose",
-        "--json-add-tracefile",
-        "coverage.json.gz",
-        "--txt-metric",
-        "decision",
+        "--json-add-tracefile=coverage.json.gz",
+        "--txt-metric=decision",
         "--txt-summary",
         "--txt=coverage.txt",
     )
@@ -67,8 +62,65 @@ def test_decisions(gcovr_test_exec: "GcovrTestExec") -> None:
 
 
 @pytest.mark.skipif(
+    not USE_PROFDATA_POSSIBLE,
+    reason="Parsing of decision is independent of OS and we do not want to have separate data for Windows and Darwin and LLVM profdata is not compatible with GCC coverage data.",
+)
+def test_decisions_llvm_profdata(gcovr_test_exec: "GcovrTestExec", check) -> None:  # type: ignore[no-untyped-def]
+    """Test of decision parsing."""
+    gcovr_test_exec.use_llvm_profdata = True
+    gcovr_test_exec.copy_source(Path("source", "decisions"))
+
+    mcdc_supported_by_compiler = gcovr_test_exec.is_in_gcc_help("coverage-mcdc")
+    additional_options = []
+    if mcdc_supported_by_compiler:
+        additional_options.append("-fcoverage-mcdc")
+
+    gcovr_test_exec.cxx_link(
+        "testcase",
+        *additional_options,
+        "main.cpp",
+        "switch_test.cpp",
+    )
+
+    gcovr_test_exec.run("./testcase")
+    process = gcovr_test_exec.gcovr(
+        "--verbose",
+        "--delete-input-files",
+        "--decisions",
+        "--llvm-cov-binary=./testcase",
+        "--json-pretty",
+        "--json=coverage.json.gz",
+        "default.profraw",
+    )
+    if mcdc_supported_by_compiler:
+        check.is_in(
+            "Found 'mcdc_records' in exported JSON report. This is ignored by GCOVR.",
+            process.stderr,
+        )
+    else:
+        check.is_not_in(
+            "Found 'mcdc_records' in exported JSON report. This is ignored by GCOVR.",
+            process.stderr,
+        )
+    gcovr_test_exec.gcovr(
+        "--verbose",
+        "--json-add-tracefile=coverage.json.gz",
+        "--json-pretty",
+        "--json=coverage.json",
+    )
+    gcovr_test_exec.gcovr(
+        "--verbose",
+        "--json-add-tracefile=coverage.json.gz",
+        "--decision",
+        "--json-summary-pretty",
+        "--json-summary=coverage_summary.json",
+    )
+    gcovr_test_exec.compare_json()
+
+
+@pytest.mark.skipif(
     not IS_LINUX,
-    reason="Parsing of decision is independent of OS and we do not want to have separate data wor Windows and Darwin.",
+    reason="Parsing of decision is independent of OS and we do not want to have separate data for Windows and Darwin.",
 )
 def test_decisions_neg_delta(gcovr_test_exec: "GcovrTestExec") -> None:
     """This test case causes a negative delta value during the multiline decision analysis, which results in a: